GCR Calibration: - now includes asymmetry - now looks @ all Z's simultaneously (although still includes some per Z plots) - now recalculates pedestals for each input file. (peds drift during flight)
Asymmetry calibration:
  - now subtracts out average asymmetry from each data point in histogram - reducing spread in each bin - average asym is added back in during fit stage.
  - now supports multiple # of slices per xtal

+++ src/Thresh/genAliveHists.cxx
@@ -1,4 +1,4 @@
-// $Header: /nfs/slac/g/glast/ground/cvs/calibGenCAL/src/Thresh/genFLEHists.cxx,v 1.5 2008/05/19 17:37:29 fewtrell Exp $
+// $Header: /nfs/slac/g/glast/ground/cvs/calibGenCAL/src/Thresh/genAliveHists.cxx,v 1.1 2008/06/25 15:15:40 chehtman Exp $
 
 /** @file
     @author Zachary Fewtrell
@@ -433,8 +433,8 @@ int main(int argc,
 	(hp2nhex8[twr])->Fill(col,lyr,p2nhex8);	
       }
       for (FaceIdx faceIdx; faceIdx.isValid(); faceIdx++){
-	int nlex8 = lex8Hists.produceHist(faceIdx).GetEntries();
-	int nhex8 = hex8Hists.produceHist(faceIdx).GetEntries();
+	int nlex8 = (int)lex8Hists.produceHist(faceIdx).GetEntries();
+	int nhex8 = (int)hex8Hists.produceHist(faceIdx).GetEntries();
 	float avlex8 = lex8Hists.produceHist(faceIdx).GetMean();
 	float avhex8 = hex8Hists.produceHist(faceIdx).GetMean();
 	if(maxnlex8 < nlex8)maxnlex8=nlex8;
